% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{plot.pwrSim_sample_size}
\alias{plot.pwrSim_sample_size}
\title{Plot Power Curve}
\usage{
\method{plot}{pwrSim_sample_size}(x, ...)
}
\arguments{
\item{x}{Object of class "pwrSim_sample_size" from calculate_sample_size()}

\item{...}{Additional arguments passed to plot}
}
\value{
NULL (creates a plot)
}
\description{
Creates a power curve plot showing the relationship between sample size
and statistical power.
}
