% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_power.R
\name{print.pwrSim_result}
\alias{print.pwrSim_result}
\title{Print Method for pwrSim Results}
\usage{
\method{print}{pwrSim_result}(x, ...)
}
\arguments{
\item{x}{Object of class "pwrSim_result"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for pwrSim Results
}
