% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{print.pwrSim_sample_size}
\alias{print.pwrSim_sample_size}
\title{Print Method for Sample Size Results}
\usage{
\method{print}{pwrSim_sample_size}(x, ...)
}
\arguments{
\item{x}{Object of class "pwrSim_sample_size"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for Sample Size Results
}
