% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_power.R
\name{summary.pwrSim_result}
\alias{summary.pwrSim_result}
\title{Summary Method for pwrSim Results}
\usage{
\method{summary}{pwrSim_result}(object, ...)
}
\arguments{
\item{object}{Object of class "pwrSim_result"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Summary Method for pwrSim Results
}
